+++ source/Examples/Resources/SqlEndpointPermission/4-RemoveConnectPermissionForTwoReplicasEachWithDifferentServiceAccount.ps1
@@ -3,6 +3,8 @@
         This example will remove connect permission to both an Always On primary replica and an
         Always On secondary replica, and where each replica has a different SQL service account.
 #>
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseDeclaredVarsMoreThanAssignments', '', Justification='The variable $ConfigurationData is used by the HQRM test')]
+param ()
 
 $ConfigurationData = @{
     AllNodes = @(
@@ -45,7 +47,7 @@ Configuration Example
 
     Import-DscResource -ModuleName 'SqlServerDsc'
 
-    node $AllNodes.Where{$_.Role -eq 'PrimaryReplica' }.NodeName
+    node $AllNodes.Where{ $_.Role -eq 'PrimaryReplica' }.NodeName
     {
         SqlEndpointPermission 'RemoveSQLConfigureEndpointPermissionPrimary'
         {